Moved validateSkill in SkillService for cleaner Controller. Added annotations for all methods from controller for rest requests.
